Flask_OCR - AI/requirements.txt
Flask==2.2.2
Flask-Caching==2.0.2
opencv-python==4.7.0.72
pytesseract==0.3.10
Pillow==9.4.0
pdf2image==1.16.2
transformers==4.27.4
torch==1.13.1
openpyxl==3.1.2
bcrypt==4.0.1